+++ src/main/java/com/fitlife/identity/service/OAuth2Service.java
@@ -1,0 +1,6 @@
+package com.fitlife.identity.service;
+import com.fitlife.identity.dto.LoginResponse;
+
+public interface OAuth2Service {
+    LoginResponse googleLogin(String token);
+}
